⚡ feat: 소스코드 구조 변경
